@@ -2,7 +2,7 @@
     "name": "p2ee/baserequirements",
     "description": "Some basic requirement types for the p2ee/preparables",
     "require-dev": {
-        "phpunit/phpunit": "4.1.*"
+        "phpunit/phpunit": "~6.3"
     },
     "license": "MIT",
     "authors": [
@@ -16,7 +16,7 @@
         "php": ">=5.4.99",
         "p2ee/preparables": "dev-master",
         "rg/injektor": ">=1.2.0,<2.0.0-dev",
-        "symfony/http-foundation": "~2.4"
+        "symfony/http-foundation": "~3.3"
     },
     "autoload": {
         "psr-0": {"p2ee\\": "src/"}
